endpoint products/:productId/sales changed
