+++ routes/sales.js
@@ -7,38 +7,6 @@ const isBefore = require('date-fns/is_before');
 const { stringToDate } = require('../utils/dates');
 const status = require('../utils/statusCodes');
 
-router.get('/products/:productId', protected, async (req, res) => {
-  try {
-    let sales;
-    if (req.params.productId) {
-      if (!req.query.from && !req.query.to) {
-        sales = await Sale.getSaleWithProduct(req.params.productId);
-        sales ? res.send(sales) : res.send({});
-      } else if (req.query.from && req.query.to) {
-        const from = stringToDate(req.query.from);
-        const to = stringToDate(req.query.to);
-        if (isBefore(from, to)) {
-          sales = await Sale.getSaleWithProductByDates(
-            req.params.productId,
-            from,
-            to
-          );
-          sales ? res.send(sales) : res.send({});
-        } else {
-          res.status(400).send('Verify dates.');
-        }
-      } else {
-        res.status(400).send('Both dates must be included.');
-      }
-    } else {
-      res.send({});
-    }
-  } catch (e) {
-    console.error(e);
-    res.sendStatus(400);
-  }
-});
-
 router.get('/', protected, async (req, res) => {
   try {
     let sales;
